requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
influxdb==5.3.1
influxdb-client==1.26.0
msgpack==1.0.3
pyserial==3.5
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
Rx==3.2.0
six==1.16.0
urllib3==1.26.8
